is_float_column moved to dataframes
