+++ deepchecks/vision/checks/data_integrity/property_label_correlation.py
@@ -20,12 +20,12 @@ from deepchecks import CheckResult, ConditionCategory, ConditionResult
 from deepchecks.core import DatasetKind
 from deepchecks.core.check_utils.feature_label_correlation_utils import get_pps_figure, pd_series_to_trace
 from deepchecks.core.errors import ModelValidationError
+from deepchecks.utils.dataframes import is_float_column
 from deepchecks.utils.strings import format_number
 from deepchecks.vision import Batch, Context, SingleDatasetCheck
 from deepchecks.vision.task_type import TaskType
 from deepchecks.vision.utils.image_properties import default_image_properties
-from deepchecks.vision.utils.property_label_correlation_utils import (calc_properties_for_property_label_correlation,
-                                                                      is_float_column)
+from deepchecks.vision.utils.property_label_correlation_utils import calc_properties_for_property_label_correlation
 
 __all__ = ['PropertyLabelCorrelation']
