+++ deepchecks/vision/utils/property_label_correlation_utils.py
@@ -11,9 +11,6 @@
 """Module for property label correlation utils."""
 from typing import List
 
-import pandas as pd
-from pandas.core.dtypes.common import is_float_dtype
-
 from deepchecks import Context
 from deepchecks.core import DatasetKind
 from deepchecks.vision.batch_wrapper import Batch
@@ -55,22 +52,3 @@ def calc_properties_for_property_label_correlation(
     data_for_properties = batch.vision_properties(imgs, image_properties, property_type)
 
     return data_for_properties, target
-
-
-def is_float_column(col: pd.Series) -> bool:
-    """Check if a column must be a float - meaning does it contain fractions.
-
-    Parameters
-    ----------
-    col : pd.Series
-        The column to check.
-
-    Returns
-    -------
-    bool
-        True if the column is float, False otherwise.
-    """
-    if not is_float_dtype(col):
-        return False
-
-    return (col.round() != col).any()
