+++ tests/utils/dataframes_test.py
@@ -10,10 +10,10 @@
 #
 """Tests for dataframes"""
 import pandas as pd
-from hamcrest import assert_that, close_to, contains_exactly
+from hamcrest import assert_that, close_to, contains_exactly, equal_to
 from scipy.stats import pearsonr
 
-from deepchecks.utils.dataframes import generalized_corrwith
+from deepchecks.utils.dataframes import generalized_corrwith, is_float_column
 
 
 def test_generalized_corrwith():
@@ -23,3 +23,20 @@ def test_generalized_corrwith():
     assert_that(result.loc['blue', 'cat'], close_to(1.0, 0.01))
     assert_that(result.index, contains_exactly('blue', 'yellow', 'red'))
     assert_that(result.columns, contains_exactly('cat', 'dog'))
+
+
+def test_is_float_column():
+    col = pd.Series([1, 2, 3, 4, 5])
+    assert_that(is_float_column(col), equal_to(False))
+
+    col = pd.Series(['a', 'b', 'c'])
+    assert_that(is_float_column(col), equal_to(False))
+
+    col = pd.Series(['a', 'b', 5.5])
+    assert_that(is_float_column(col), equal_to(False))
+
+    col = pd.Series([1, 2, 3, 4, 5], dtype='float')
+    assert_that(is_float_column(col), equal_to(False))
+
+    col = pd.Series([1, 2, 3, 4, 5.5], dtype='float64')
+    assert_that(is_float_column(col), equal_to(True))
